Now, the shadow map output texture index is 0. Only using 1 fbo and 2 textures for shadow mapping now. Removed SHADOW_MAP_GENERATION_TEXTURE_UNIT, as it wasn't needed. Replace a use_texture call with set_sampler_texture_unit_for_shader. Find some bugs in the shadow map generation process (not in the shadowing algorithm, but in the framebuffer ping pong process. Fix a type precision error in utils.c.

+++ src/audio_visual/accel/src/utils.c
@@ -180,7 +180,7 @@ void loop_application(const Screen* const screen, void (*const drawer) (const St
 		#ifndef USE_VSYNC
 		const GLfloat ms_elapsed = (GLfloat) (SDL_GetPerformanceCounter() - before) * one_over_performance_freq * 1000.0f;
 		const GLfloat wait_for_exact_fps = max_delay - ms_elapsed;
-		if (wait_for_exact_fps > 12.0f) SDL_Delay(wait_for_exact_fps - 0.5f); // SDL_Delay tends to be late, so 0.5f accounts for that
+		if (wait_for_exact_fps > 12.0f) SDL_Delay((Uint32) (wait_for_exact_fps - 0.5f)); // SDL_Delay tends to be late, so 0.5f accounts for that
 		#endif
 	}
